be more thorough when rendering not 'above_vertices'
Specifically, determine the index of the lowest vertex
when creating a traveller that must be below it... because
other things may be painted (e.g., the river in Prim's)
which spoil it.

This is the wrong way to do it: better to have vertices
and edges in different containers, and then have travellers
themselve in containers above or below the vertices' one.